Refactored generators, improved RNG selection formula

+++ Source/ACE.Database/Models/World/WeenieExtensions.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Linq;
 using ACE.Database.Models.Shard;
 using ACE.Entity;
@@ -363,8 +362,8 @@ namespace ACE.Database.Models.World
                     Probability = value.Probability,
                     WeenieClassId = value.WeenieClassId,
                     Delay = value.Delay,
-                    InitCreate = value.InitCreate,
-                    MaxCreate = value.MaxCreate,
+                    InitCreate = (int)value.InitCreate,     // FIXME in SQL generator: should be int?
+                    MaxCreate = (int)value.MaxCreate,
                     WhenCreate = value.WhenCreate,
                     WhereCreate = value.WhereCreate,
                     StackSize = value.StackSize,
